# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v2/events": {
            "get": {
                "description": "Returns a list of events with details, including category, location, audience, and organizer info. Supports historical filtering, limit, and offset.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get events",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include historical events",
                        "name": "historical",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Event"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Creates a new event with the provided details. Requires authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Create a new event",
                "parameters": [
                    {
                        "description": "Event to create",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NewEvent"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v2/events/audiences": {
            "get": {
                "description": "Retrieves a list of all event audiences.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get event audiences",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v2/events/categories": {
            "get": {
                "description": "Retrieves a list of all event categories active.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get event categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v2/events/time": {
            "get": {
                "description": "Retrieves a list of all time types for events.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get all time types",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v2/events/{id}": {
            "get": {
                "description": "Retrieves a specific protected event by its ID, including detailed information such as category, location, audience, and organizer info.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get protected event by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Updates an existing event with the provided details. Requires authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Update an existing event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated event details",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NewEvent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Deletes an event by its ID. Requires authentication.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Delete an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal.LocalTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        },
...
        "models.Audience": {
            "type": "object",
            "required": [
                "name_en",
                "name_no"
            ],
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "id": {
                    "type": "integer"
                },
                "name_en": {
                    "type": "string"
                },
                "name_no": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                }
            }
        },
        "models.Category": {
            "type": "object",
            "required": [
                "color",
                "name_en",
                "name_no"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "id": {
                    "type": "integer"
                },
                "name_en": {
                    "type": "string"
                },
                "name_no": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                }
            }
        },
        "models.Cities": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.Event": {
            "type": "object",
            "required": [
                "description_en",
                "description_no",
                "name_en",
                "name_no",
                "time_end",
                "time_publish",
                "time_start"
            ],
            "properties": {
                "audience": {
                    "$ref": "#/definitions/models.Audience"
                },
                "canceled": {
                    "type": "boolean"
                },
                "capacity": {
                    "type": "integer"
                },
                "category": {
                    "$ref": "#/definitions/models.Category"
                },
                "created_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "description_en": {
                    "type": "string"
                },
                "description_no": {
                    "type": "string"
                },
                "digital": {
                    "type": "boolean"
                },
                "highlight": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "image_banner": {
                    "type": "string"
                },
                "image_small": {
                    "type": "string"
                },
                "informational_en": {
                    "type": "string"
                },
                "informational_no": {
                    "type": "string"
                },
                "is_full": {
                    "type": "boolean"
                },
                "link_discord": {
                    "type": "string"
                },
                "link_facebook": {
                    "type": "string"
                },
                "link_signup": {
                    "type": "string"
                },
                "link_stream": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/models.Location"
                },
                "name_en": {
                    "type": "string"
                },
                "name_no": {
                    "type": "string"
                },
                "organization": {
                    "$ref": "#/definitions/models.Organization"
                },
                "parent_id": {
                    "type": "integer"
                },
                "rule": {
                    "$ref": "#/definitions/models.Rule"
                },
                "time_end": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "time_publish": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "time_signup_deadline": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "time_signup_release": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "time_start": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "time_type": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "visible": {
                    "type": "boolean"
                }
            }
        },
...
        "models.Location": {
            "type": "object",
            "required": [
                "name_en",
                "name_no"
            ],
            "properties": {
                "address_postcode": {
                    "type": "integer"
                },
                "address_street": {
                    "type": "string"
                },
                "cities": {
                    "$ref": "#/definitions/models.Cities"
                },
                "coordinate_lat": {
                    "type": "number",
                    "maximum": 90,
                    "minimum": -90
                },
                "coordinate_lon": {
                    "type": "number",
                    "maximum": 180,
                    "minimum": -180
                },
                "created_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "id": {
                    "type": "integer"
                },
                "mazemap_campus_id": {
                    "type": "integer"
                },
                "mazemap_poi_id": {
                    "type": "integer"
                },
                "name_en": {
                    "type": "string"
                },
                "name_no": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "models.NewEvent": {
            "type": "object",
            "required": [
                "category_id",
                "description_en",
                "description_no",
                "name_en",
                "name_no",
                "time_end",
                "time_publish",
                "time_start"
            ],
            "properties": {
                "audience_id": {
                    "type": "integer"
                },
                "canceled": {
                    "type": "boolean"
                },
                "capacity": {
                    "type": "integer"
                },
                "category_id": {
                    "type": "integer"
                },
                "created_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "description_en": {
                    "type": "string"
                },
                "description_no": {
                    "type": "string"
                },
                "digital": {
                    "type": "boolean"
                },
                "highlight": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "image_banner": {
                    "type": "string"
                },
                "image_small": {
                    "type": "string"
                },
                "informational_en": {
                    "type": "string"
                },
                "informational_no": {
                    "type": "string"
                },
                "is_full": {
                    "type": "boolean"
                },
                "link_discord": {
                    "type": "string"
                },
                "link_facebook": {
                    "type": "string"
                },
                "link_signup": {
                    "type": "string"
                },
                "link_stream": {
                    "type": "string"
                },
                "location_id": {
                    "type": "integer"
                },
                "name_en": {
                    "type": "string"
                },
                "name_no": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "integer"
                },
                "parent_id": {
                    "type": "integer"
                },
                "rule_id": {
                    "type": "integer"
                },
                "time_end": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "time_publish": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "time_signup_deadline": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "time_signup_release": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "time_start": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "time_type": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "visible": {
                    "type": "boolean"
                }
            }
        },
...
        "models.Organization": {
            "type": "object",
            "required": [
                "description_en",
                "description_no",
                "link_homepage",
                "name_en",
                "name_no"
            ],
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "description_en": {
                    "type": "string"
                },
                "description_no": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "link_facebook": {
                    "type": "string"
                },
                "link_homepage": {
                    "type": "string"
                },
                "link_instagram": {
                    "type": "string"
                },
                "link_linkedin": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name_en": {
                    "type": "string"
                },
                "name_no": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                }
            }
        },
...
        "models.Rule": {
            "type": "object",
            "required": [
                "description_en",
                "description_no",
                "name_en",
                "name_no"
            ],
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                },
                "description_en": {
                    "type": "string"
                },
                "description_no": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name_en": {
                    "type": "string"
                },
                "name_no": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/internal.LocalTime"
                }
            }
        },
...
    }
}